rename samplers to optimizers
