+++ src/hyperactive/opt/optuna/__init__.py
@@ -1,22 +1,22 @@
 """Individual Optuna optimization algorithms."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
-from ._cmaes_sampler import CmaEsSampler
-from ._gp_sampler import GPSampler
-from ._grid_sampler import GridSampler
-from ._nsga_ii_sampler import NSGAIISampler
-from ._nsga_iii_sampler import NSGAIIISampler
-from ._qmc_sampler import QMCSampler
-from ._random_sampler import RandomSampler
-from ._tpe_sampler import TPESampler
+from ._cmaes_optimizer import CmaEsOptimizer
+from ._gp_optimizer import GPOptimizer
+from ._grid_optimizer import GridOptimizer
+from ._nsga_ii_optimizer import NSGAIIOptimizer
+from ._nsga_iii_optimizer import NSGAIIIOptimizer
+from ._qmc_optimizer import QMCOptimizer
+from ._random_optimizer import RandomOptimizer
+from ._tpe_optimizer import TPEOptimizer
 
 __all__ = [
-    "TPESampler",
-    "RandomSampler",
-    "CmaEsSampler",
-    "GPSampler",
-    "GridSampler",
-    "NSGAIISampler",
-    "NSGAIIISampler",
-    "QMCSampler",
+    "TPEOptimizer",
+    "RandomOptimizer",
+    "CmaEsOptimizer",
+    "GPOptimizer",
+    "GridOptimizer",
+    "NSGAIIOptimizer",
+    "NSGAIIIOptimizer",
+    "QMCOptimizer",
 ]
