+++ src/hyperactive/opt/optuna/_nsga_iii_optimizer.py
@@ -1,11 +1,11 @@
-"""Random sampler optimizer."""
+"""NSGA-III multi-objective optimizer."""
 # copyright: hyperactive developers, MIT License (see LICENSE file)
 
 from .._adapters._base_optuna_adapter import _BaseOptunaAdapter
 
 
-class RandomSampler(_BaseOptunaAdapter):
-    """Random sampler optimizer.
+class NSGAIIIOptimizer(_BaseOptunaAdapter):
+    """NSGA-III multi-objective optimizer.
 
     Parameters
     ----------
@@ -26,16 +26,22 @@ class RandomSampler(_BaseOptunaAdapter):
         Number of trials after which to stop if no improvement.
     max_score : float, default=None
         Maximum score threshold. Stop optimization when reached.
+    population_size : int, default=50
+        Population size for NSGA-III.
+    mutation_prob : float, default=0.1
+        Mutation probability for NSGA-III.
+    crossover_prob : float, default=0.9
+        Crossover probability for NSGA-III.
     experiment : BaseExperiment, optional
         The experiment to optimize parameters for.
         Optional, can be passed later via ``set_params``.
 
     Examples
     --------
-    Basic usage of RandomSampler with a scikit-learn experiment:
+    Basic usage of NSGAIIIOptimizer with a scikit-learn experiment:
 
     >>> from hyperactive.experiment.integrations import SklearnCvExperiment
-    >>> from hyperactive.opt.optuna import RandomSampler
+    >>> from hyperactive.opt.optuna import NSGAIIIOptimizer
     >>> from sklearn.datasets import load_iris
     >>> from sklearn.svm import SVC
     >>> X, y = load_iris(return_X_y=True)
@@ -44,17 +50,17 @@ class RandomSampler(_BaseOptunaAdapter):
     ...     "C": (0.01, 10),
     ...     "gamma": (0.0001, 10),
     ... }
-    >>> optimizer = RandomSampler(
+    >>> optimizer = NSGAIIIOptimizer(
     ...     param_space=param_space, n_trials=50, experiment=sklearn_exp
     ... )
     >>> best_params = optimizer.run()
     """
 
     _tags = {
-        "info:name": "Random Sampler",
+        "info:name": "NSGA-III Optimizer",
         "info:local_vs_global": "global",
-        "info:explore_vs_exploit": "explore",
-        "info:compute": "low",
+        "info:explore_vs_exploit": "mixed",
+        "info:compute": "high",
         "python_dependencies": ["optuna"],
     }
 
@@ -66,8 +72,15 @@ class RandomSampler(_BaseOptunaAdapter):
         random_state=None,
         early_stopping=None,
         max_score=None,
+        population_size=50,
+        mutation_prob=0.1,
+        crossover_prob=0.9,
         experiment=None,
     ):
+        self.population_size = population_size
+        self.mutation_prob = mutation_prob
+        self.crossover_prob = crossover_prob
+
         super().__init__(
             param_space=param_space,
             n_trials=n_trials,
@@ -78,18 +91,36 @@ class RandomSampler(_BaseOptunaAdapter):
             experiment=experiment,
         )
 
-    def _get_sampler(self):
-        """Get the Random sampler.
+    def _get_optimizer(self):
+        """Get the NSGA-III optimizer.
 
         Returns
         -------
-        sampler
-            The Optuna RandomSampler instance
+        optimizer
+            The Optuna NSGAIIIOptimizer instance
         """
         import optuna
 
-        sampler_kwargs = {}
-        if self.random_state is not None:
-            sampler_kwargs["seed"] = self.random_state
+        optimizer_kwargs = {
+            "population_size": self.population_size,
+            "mutation_prob": self.mutation_prob,
+            "crossover_prob": self.crossover_prob,
+        }
 
-        return optuna.samplers.RandomSampler(**sampler_kwargs)
+        if self.random_state is not None:
+            optimizer_kwargs["seed"] = self.random_state
+
+        return optuna.samplers.NSGAIIISampler(**optimizer_kwargs)
+
+    @classmethod
+    def get_test_params(cls, parameter_set="default"):
+        """Return testing parameter settings for the optimizer."""
+        params = super().get_test_params(parameter_set)
+        params[0].update(
+            {
+                "population_size": 20,
+                "mutation_prob": 0.2,
+                "crossover_prob": 0.8,
+            }
+        )
+        return params
